responsabilidades da classe de serviço concluída
