+++ jogo-oito/src/main/java/chat/gpt/builder/GameStart.java
@@ -2,14 +2,16 @@ package chat.gpt.builder;
 
 import chat.gpt.controller.GameController;
 import chat.gpt.controller.GameService;
+import chat.gpt.model.Grid;
 import chat.gpt.view.GameGUI;
 
 public class GameStart {
 
     public static void startGame() {
-        GameService gameService = new GameService();
-        GameController gameController = new GameController(gameService, null);
-        GameGUI gameGUI = new GameGUI(gameService, gameController);
+        Grid grid = new Grid();
+        GameService gameService = new GameService(grid);
+        GameController gameController = new GameController(gameService, null, grid);
+        GameGUI gameGUI = new GameGUI(gameController);
         gameController.setView(gameGUI);
     }
